PyCSDL2_RWread(): cast signed buf.len to unsigned size_t
Fix the signed-unsigned integer comparisons by casting the Py_ssize_t
buf.len to size_t. There should not be any overflow since it should not
be negative.

+++ src/rwops.h
@@ -630,7 +630,7 @@ PyCSDL2_RWread(PyObject *self, PyObject *args, PyObject *kwds)
                                      &maxnum))
         return NULL;
 
-    if (buf.len != size * maxnum) {
+    if ((size_t)buf.len != size * maxnum) {
         PyBuffer_Release(&buf);
         return PyErr_Format(PyExc_BufferError, "Invalid buffer size");
     }
